Split DefaultPage into FirstPartyRequestPage and ThirdPartyRequestPage to eliminate conditional branching and improve maintainability
