+++ src/applications/claims-status/containers/DocumentRequestPage.jsx
@@ -7,7 +7,8 @@ import { scrollToTop } from 'platform/utilities/scroll';
 import NeedHelp from '../components/NeedHelp';
 import ClaimsBreadcrumbs from '../components/ClaimsBreadcrumbs';
 import Notification from '../components/Notification';
-import DefaultPage from '../components/claim-document-request-pages/DefaultPage';
+import FirstPartyRequestPage from '../components/claim-document-request-pages/FirstPartyRequestPage';
+import ThirdPartyRequestPage from '../components/claim-document-request-pages/ThirdPartyRequestPage';
 import {
   cancelUpload,
   clearNotification,
@@ -62,36 +63,36 @@ class DocumentRequestPage extends React.Component {
     this.props.navigate(redirectPath);
   }
 
-  getDefaultPage() {
+  getRequestPage() {
     const {
       message,
       type1UnknownErrors,
       timezoneMitigationEnabled,
       showDocumentUploadStatus,
     } = this.props;
-    return (
-      <>
-        <DefaultPage
-          item={this.props.trackedItem}
-          message={showDocumentUploadStatus ? message : null}
-          onCancel={this.props.cancelUpload}
-          onSubmit={files =>
-            this.props.submitFiles(
-              this.props.claim.id,
-              this.props.trackedItem,
-              files,
-              showDocumentUploadStatus,
-              timezoneMitigationEnabled,
-            )
-          }
-          progress={this.props.progress}
-          type1UnknownErrors={
-            showDocumentUploadStatus ? type1UnknownErrors : null
-          }
-          uploading={this.props.uploading}
-        />
-      </>
-    );
+
+    const pageProps = {
+      item: this.props.trackedItem,
+      message: showDocumentUploadStatus ? message : null,
+      onCancel: this.props.cancelUpload,
+      onSubmit: files =>
+        this.props.submitFiles(
+          this.props.claim.id,
+          this.props.trackedItem,
+          files,
+          showDocumentUploadStatus,
+          timezoneMitigationEnabled,
+        ),
+      progress: this.props.progress,
+      type1UnknownErrors: showDocumentUploadStatus ? type1UnknownErrors : null,
+      uploading: this.props.uploading,
+    };
+
+    if (this.props.trackedItem?.status === 'NEEDED_FROM_YOU') {
+      return <FirstPartyRequestPage {...pageProps} />;
+    }
+
+    return <ThirdPartyRequestPage {...pageProps} />;
   }
 
   render() {
@@ -162,7 +163,7 @@ class DocumentRequestPage extends React.Component {
           {TrackedItem.isAutomated5103Notice(trackedItem.displayName) ? (
             <Default5103EvidenceNotice item={trackedItem} />
           ) : (
-            <>{this.getDefaultPage()}</>
+            <>{this.getRequestPage()}</>
           )}
         </>
       );
